+++ app/src/main/AndroidManifest.xml
@@ -9,8 +9,8 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".gui.LoginActivity"></activity>
-        <activity android:name=".gui.MainActivity">
+        <activity android:name=".infra.gui.LoginActivity"></activity>
+        <activity android:name=".infra.gui.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
