+++ app/src/main/java/br/ufrpe/bsi/mpoo/petSpeed/animal/persistencia/AnimalDAO.java
@@ -1,0 +1,46 @@
+package br.ufrpe.bsi.mpoo.petSpeed.animal.persistencia;
+
+
+import java.util.List;
+
+import br.ufrpe.bsi.mpoo.petSpeed.animal.dominio.Animal;
+import br.ufrpe.bsi.mpoo.petSpeed.os.dominio.OrdemServico;
+import br.ufrpe.bsi.mpoo.petSpeed.infra.DBHelper;
+
+public class AnimalDAO {
+
+	private DBHelper db;
+
+	public void cadastraAnimal(Animal animal) {
+
+	}
+
+	public void deletaAnimal() {
+
+	}
+
+	public void alteraFoto() {
+
+	}
+
+	public void adicionaOS() {
+
+	}
+
+	public Animal getAnimalById() {
+		return null;
+	}
+
+	public OrdemServico getHistoricoById() {
+		return null;
+	}
+
+	public List<OrdemServico> getHistoricoByMedico() {
+		return null;
+	}
+
+	public List<OrdemServico> getAllHistorico() {
+		return null;
+	}
+
+}
