+++ app/src/main/java/br/ufrpe/bsi/mpoo/petSpeed/cliente/persistencia/ClienteDAO.java
@@ -1,0 +1,71 @@
+package br.ufrpe.bsi.mpoo.petSpeed.cliente.persistencia;
+
+
+import java.util.List;
+
+import br.ufrpe.bsi.mpoo.petSpeed.animal.dominio.Animal;
+import br.ufrpe.bsi.mpoo.petSpeed.cliente.dominio.Cliente;
+import br.ufrpe.bsi.mpoo.petSpeed.pessoa.dominio.Endereco;
+import br.ufrpe.bsi.mpoo.petSpeed.infra.DBHelper;
+
+public class ClienteDAO {
+
+	private DBHelper db;
+
+	public void cadastraCliente() {
+
+	}
+
+	public void deletaCliente() {
+
+	}
+
+	public Endereco getEnderecoById() {
+		return null;
+	}
+
+	public Animal getAnimalById() {
+		return null;
+	}
+
+	public Animal getAnimalByRaca() {
+		return null;
+	}
+
+	public List<Endereco> getAllEndereco() {
+		return null;
+	}
+
+	public List<Animal> getAllAnimal() {
+		return null;
+	}
+
+	public void removeEndereco() {
+
+	}
+
+	public void adicionaEndereco() {
+
+	}
+
+	public Cliente getClienteById() {
+		return null;
+	}
+
+	public Cliente getClienteByEmail() {
+		return null;
+	}
+
+	public void alteraEmail() {
+
+	}
+
+	public void alteraSenha() {
+
+	}
+
+	public void alteraAvaliacao() {
+
+	}
+
+}
