+++ app/src/main/java/br/ufrpe/bsi/mpoo/petSpeed/clinica/negocio/ClinicaServices.java
@@ -1,0 +1,55 @@
+package br.ufrpe.bsi.mpoo.petSpeed.clinica.negocio;
+
+
+import java.util.List;
+
+import br.ufrpe.bsi.mpoo.petSpeed.medico.dominio.Medico;
+import br.ufrpe.bsi.mpoo.petSpeed.clinica.persistencia.ClinicaDAO;
+import br.ufrpe.bsi.mpoo.petSpeed.pessoa.persistencia.EnderecoDAO;
+import br.ufrpe.bsi.mpoo.petSpeed.medico.persistencia.MedicoDAO;
+
+public class ClinicaServices {
+
+	private ClinicaDAO clinicaDAO;
+
+	private MedicoDAO medicoDAO;
+
+	private EnderecoDAO enderecoDAO;
+
+	public void cadastraClinica() {
+
+	}
+
+	public void deletaClinica() {
+
+	}
+
+	public void alteraAvaliacao() {
+
+	}
+
+	public void alteraEndereco() {
+
+	}
+
+	public void adicionaMedico() {
+
+	}
+
+	public void removeMedico() {
+
+	}
+
+	public void alteraCrmv() {
+
+	}
+
+	public Medico getMedicoById() {
+		return null;
+	}
+
+	public List<Medico> getAllMedico() {
+		return null;
+	}
+
+}
