--- app/src/main/java/br/ufrpe/bsi/mpoo/petSpeed/dominio/Endereco.java
@@ -1,46 +1,0 @@
-package br.ufrpe.bsi.mpoo.petSpeed.dominio;
-
-public class Endereco {
-
-	private long id;
-
-	private String cep;
-
-	private long numero;
-
-	private String complemento;
-
-
-
-	public long getId() {
-		return id;
-	}
-
-	public void setId(long id) {
-		this.id = id;
-	}
-
-	public String getCep() {
-		return cep;
-	}
-
-	public void setCep(String cep) {
-		this.cep = cep;
-	}
-
-	public long getNumero() {
-		return numero;
-	}
-
-	public void setNumero(long numero) {
-		this.numero = numero;
-	}
-
-	public String getComplemento() {
-		return complemento;
-	}
-
-	public void setComplemento(String complemento) {
-		this.complemento = complemento;
-	}
-}
