--- app/src/main/java/br/ufrpe/bsi/mpoo/petSpeed/dominio/Medico.java
@@ -1,88 +1,0 @@
-package br.ufrpe.bsi.mpoo.petSpeed.dominio;
-
-import java.util.List;
-
-public class Medico {
-
-	private long id;
-
-	private String email;
-
-	private String senha;
-
-	private double avaliacao;
-
-	private String crmv;
-
-	private Endereco endereco;
-
-	private Pessoa dadosPessoais;
-
-	private List<Clinica> clinicas;
-
-
-
-	public long getId() {
-		return id;
-	}
-
-	public void setId(long id) {
-		this.id = id;
-	}
-
-	public String getEmail() {
-		return email;
-	}
-
-	public void setEmail(String email) {
-		this.email = email;
-	}
-
-	public String getSenha() {
-		return senha;
-	}
-
-	public void setSenha(String senha) {
-		this.senha = senha;
-	}
-
-	public double getAvaliacao() {
-		return avaliacao;
-	}
-
-	public void setAvaliacao(double avaliacao) {
-		this.avaliacao = avaliacao;
-	}
-
-	public String getCrmv() {
-		return crmv;
-	}
-
-	public void setCrmv(String crmv) {
-		this.crmv = crmv;
-	}
-
-	public Endereco getEndereco() {
-		return endereco;
-	}
-
-	public void setEndereco(Endereco endereco) {
-		this.endereco = endereco;
-	}
-
-	public Pessoa getDadosPessoais() {
-		return dadosPessoais;
-	}
-
-	public void setDadosPessoais(Pessoa dadosPessoais) {
-		this.dadosPessoais = dadosPessoais;
-	}
-
-	public List<Clinica> getClinicas() {
-		return clinicas;
-	}
-
-	public void setClinicas(List<Clinica> clinicas) {
-		this.clinicas = clinicas;
-	}
-}
