--- app/src/main/java/br/ufrpe/bsi/mpoo/petSpeed/dominio/OrdemServico.java
@@ -1,76 +1,0 @@
-package br.ufrpe.bsi.mpoo.petSpeed.dominio;
-
-public class OrdemServico {
-
-	private long id;
-
-	private Medico medico;
-
-	private Triagem triagem;
-
-	private Cliente cliente;
-
-	private Animal animal;
-
-	private String descricao;
-
-	private Prioridade prioridade;
-
-
-
-	public long getId() {
-		return id;
-	}
-
-	public void setId(long id) {
-		this.id = id;
-	}
-
-	public Medico getMedico() {
-		return medico;
-	}
-
-	public void setMedico(Medico medico) {
-		this.medico = medico;
-	}
-
-	public Triagem getTriagem() {
-		return triagem;
-	}
-
-	public void setTriagem(Triagem triagem) {
-		this.triagem = triagem;
-	}
-
-	public Cliente getCliente() {
-		return cliente;
-	}
-
-	public void setCliente(Cliente cliente) {
-		this.cliente = cliente;
-	}
-
-	public Animal getAnimal() {
-		return animal;
-	}
-
-	public void setAnimal(Animal animal) {
-		this.animal = animal;
-	}
-
-	public String getDescricao() {
-		return descricao;
-	}
-
-	public void setDescricao(String descricao) {
-		this.descricao = descricao;
-	}
-
-	public Prioridade getPrioridade() {
-		return prioridade;
-	}
-
-	public void setPrioridade(Prioridade prioridade) {
-		this.prioridade = prioridade;
-	}
-}
