--- app/src/main/java/br/ufrpe/bsi/mpoo/petSpeed/dominio/Prioridade.java
@@ -1,9 +1,0 @@
-package br.ufrpe.bsi.mpoo.petSpeed.dominio;
-
-public enum Prioridade{
-
-	ALTA,
-
-	BAIXA
-
-}
