--- app/src/main/java/br/ufrpe/bsi/mpoo/petSpeed/gui/LoginActivity.java
@@ -1,15 +1,0 @@
-package br.ufrpe.bsi.mpoo.petSpeed.gui;
-
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-
-import br.ufrpe.bsi.mpoo.petSpeed.R;
-
-public class LoginActivity extends AppCompatActivity {
-
-	@Override
-	protected void onCreate(Bundle savedInstanceState) {
-		super.onCreate(savedInstanceState);
-		setContentView(R.layout.activity_login);
-	}
-}
