+++ app/src/main/java/br/ufrpe/bsi/mpoo/petSpeed/infra/negocio/TriagemServices.java
@@ -1,0 +1,5 @@
+package br.ufrpe.bsi.mpoo.petSpeed.infra.negocio;
+
+public class TriagemServices {
+
+}
