+++ app/src/main/java/br/ufrpe/bsi/mpoo/petSpeed/infra/negocio/slopeOne.java
@@ -1,0 +1,9 @@
+package br.ufrpe.bsi.mpoo.petSpeed.infra.negocio;
+
+public class slopeOne {
+
+	public void calculaRecomendaçao() {
+
+	}
+
+}
