--- app/src/main/java/br/ufrpe/bsi/mpoo/petSpeed/infra/persistencia/ClinicaDAO.java
@@ -1,49 +1,0 @@
-package br.ufrpe.bsi.mpoo.petSpeed.infra.persistencia;
-
-
-import java.util.List;
-
-import br.ufrpe.bsi.mpoo.petSpeed.dominio.Medico;
-import br.ufrpe.bsi.mpoo.petSpeed.infra.DBHelper;
-
-public class ClinicaDAO {
-
-	private DBHelper db;
-
-	public void cadastraClinica() {
-
-	}
-
-	public void deletaClinica() {
-
-	}
-
-	public void alteraAvaliacao() {
-
-	}
-
-	public void alteraEndereco() {
-
-	}
-
-	public void adicionaMedico() {
-
-	}
-
-	public void removeMedico() {
-
-	}
-
-	public void alteraCrmv() {
-
-	}
-
-	public Medico getMedicoById() {
-		return null;
-	}
-
-	public List<Medico> getAllMedico() {
-		return null;
-	}
-
-}
