--- app/src/main/java/br/ufrpe/bsi/mpoo/petSpeed/infra/persistencia/EnderecoDAO.java
@@ -1,35 +1,0 @@
-package br.ufrpe.bsi.mpoo.petSpeed.infra.persistencia;
-
-
-import br.ufrpe.bsi.mpoo.petSpeed.dominio.Endereco;
-import br.ufrpe.bsi.mpoo.petSpeed.infra.DBHelper;
-
-public class EnderecoDAO {
-
-	private DBHelper db;
-
-	public void cadastraEndereco() {
-
-	}
-
-	public void deletaEndereco() {
-
-	}
-
-	public void alteraCep() {
-
-	}
-
-	public void alteraNumero() {
-
-	}
-
-	public void alteraComplemento() {
-
-	}
-
-	public Endereco getEnderecoById() {
-		return null;
-	}
-
-}
