--- app/src/main/java/br/ufrpe/bsi/mpoo/petSpeed/infra/persistencia/OrdemServicoDAO.java
@@ -1,33 +1,0 @@
-package br.ufrpe.bsi.mpoo.petSpeed.infra.persistencia;
-
-
-import java.util.List;
-
-import br.ufrpe.bsi.mpoo.petSpeed.dominio.OrdemServico;
-import br.ufrpe.bsi.mpoo.petSpeed.infra.DBHelper;
-
-public class OrdemServicoDAO {
-
-	private DBHelper db;
-
-	public void cadastraOS() {
-
-	}
-
-	public void deletaOS() {
-
-	}
-
-	public OrdemServico GetOSbyId() {
-		return null;
-	}
-
-	public List<OrdemServico> getOSbyPrioridade() {
-		return null;
-	}
-
-	public List<OrdemServico> getAllOS() {
-		return null;
-	}
-
-}
