+++ app/src/main/java/br/ufrpe/bsi/mpoo/petSpeed/medico/dominio/Medico.java
@@ -1,0 +1,92 @@
+package br.ufrpe.bsi.mpoo.petSpeed.medico.dominio;
+
+import java.util.List;
+
+import br.ufrpe.bsi.mpoo.petSpeed.clinica.dominio.Clinica;
+import br.ufrpe.bsi.mpoo.petSpeed.pessoa.dominio.Pessoa;
+import br.ufrpe.bsi.mpoo.petSpeed.pessoa.dominio.Endereco;
+
+public class Medico {
+
+	private long id;
+
+	private String email;
+
+	private String senha;
+
+	private double avaliacao;
+
+	private String crmv;
+
+	private Endereco endereco;
+
+	private Pessoa dadosPessoais;
+
+	private List<Clinica> clinicas;
+
+
+
+	public long getId() {
+		return id;
+	}
+
+	public void setId(long id) {
+		this.id = id;
+	}
+
+	public String getEmail() {
+		return email;
+	}
+
+	public void setEmail(String email) {
+		this.email = email;
+	}
+
+	public String getSenha() {
+		return senha;
+	}
+
+	public void setSenha(String senha) {
+		this.senha = senha;
+	}
+
+	public double getAvaliacao() {
+		return avaliacao;
+	}
+
+	public void setAvaliacao(double avaliacao) {
+		this.avaliacao = avaliacao;
+	}
+
+	public String getCrmv() {
+		return crmv;
+	}
+
+	public void setCrmv(String crmv) {
+		this.crmv = crmv;
+	}
+
+	public Endereco getEndereco() {
+		return endereco;
+	}
+
+	public void setEndereco(Endereco endereco) {
+		this.endereco = endereco;
+	}
+
+	public Pessoa getDadosPessoais() {
+		return dadosPessoais;
+	}
+
+	public void setDadosPessoais(Pessoa dadosPessoais) {
+		this.dadosPessoais = dadosPessoais;
+	}
+
+	public List<Clinica> getClinicas() {
+		return clinicas;
+	}
+
+	public void setClinicas(List<Clinica> clinicas) {
+		this.clinicas = clinicas;
+	}
+}
