+++ app/src/main/java/br/ufrpe/bsi/mpoo/petSpeed/medico/negocio/MedicoServices.java
@@ -1,0 +1,48 @@
+package br.ufrpe.bsi.mpoo.petSpeed.medico.negocio;
+
+import br.ufrpe.bsi.mpoo.petSpeed.pessoa.dominio.Endereco;
+import br.ufrpe.bsi.mpoo.petSpeed.medico.dominio.Medico;
+import br.ufrpe.bsi.mpoo.petSpeed.clinica.persistencia.ClinicaDAO;
+import br.ufrpe.bsi.mpoo.petSpeed.pessoa.persistencia.EnderecoDAO;
+import br.ufrpe.bsi.mpoo.petSpeed.medico.persistencia.MedicoDAO;
+
+public class MedicoServices {
+
+	private MedicoDAO medicoDAO;
+
+	private EnderecoDAO enderecoDAO;
+
+	private ClinicaDAO clinicaDAO;
+
+
+	public void cadastraMedico(Medico medico) {
+
+		
+
+	}
+
+	public void deletaMedico() {
+
+	}
+
+	public void alteraAvaliacao() {
+
+	}
+
+	public void alteraClinica() {
+
+	}
+
+	public void alteraEndereco() {
+
+	}
+
+	public void alteraCrmv() {
+
+	}
+
+	public Endereco getEnderecoById() {
+		return null;
+	}
+
+}
