--- app/src/main/java/br/ufrpe/bsi/mpoo/petSpeed/negocio/ClienteServices.java
@@ -1,76 +1,0 @@
-package br.ufrpe.bsi.mpoo.petSpeed.negocio;
-
-import java.util.List;
-
-import br.ufrpe.bsi.mpoo.petSpeed.dominio.Animal;
-import br.ufrpe.bsi.mpoo.petSpeed.dominio.Cliente;
-import br.ufrpe.bsi.mpoo.petSpeed.dominio.Endereco;
-import br.ufrpe.bsi.mpoo.petSpeed.infra.persistencia.AnimalDAO;
-import br.ufrpe.bsi.mpoo.petSpeed.infra.persistencia.ClienteDAO;
-import br.ufrpe.bsi.mpoo.petSpeed.infra.persistencia.EnderecoDAO;
-
-public class ClienteServices {
-
-	private ClienteDAO clienteDAO;
-
-	private EnderecoDAO enderecoDAO;
-
-	private AnimalDAO animalDAO;
-
-	public void cadastraCliente() {
-
-	}
-
-	public void deletaCliente() {
-
-	}
-
-	public Endereco getEnderecoById() {
-		return null;
-	}
-
-	public Animal getAnimalById() {
-		return null;
-	}
-
-	public Animal getAnimalByRaca() {
-		return null;
-	}
-
-	public List<Endereco> getAllEndereco() {
-		return null;
-	}
-
-	public List<Animal> getAllAnimal() {
-		return null;
-	}
-
-	public void removeEndereco() {
-
-	}
-
-	public void adicionaEndereco() {
-
-	}
-
-	public Cliente getClienteById() {
-		return null;
-	}
-
-	public Cliente getClienteByEmail() {
-		return null;
-	}
-
-	public void alteraEmail() {
-
-	}
-
-	public void alteraSenha() {
-
-	}
-
-	public void alteraAvaliacao() {
-
-	}
-
-}
