--- app/src/main/java/br/ufrpe/bsi/mpoo/petSpeed/negocio/ClinicaServices.java
@@ -1,55 +1,0 @@
-package br.ufrpe.bsi.mpoo.petSpeed.negocio;
-
-
-import java.util.List;
-
-import br.ufrpe.bsi.mpoo.petSpeed.dominio.Medico;
-import br.ufrpe.bsi.mpoo.petSpeed.infra.persistencia.ClinicaDAO;
-import br.ufrpe.bsi.mpoo.petSpeed.infra.persistencia.EnderecoDAO;
-import br.ufrpe.bsi.mpoo.petSpeed.infra.persistencia.MedicoDAO;
-
-public class ClinicaServices {
-
-	private ClinicaDAO clinicaDAO;
-
-	private MedicoDAO medicoDAO;
-
-	private EnderecoDAO enderecoDAO;
-
-	public void cadastraClinica() {
-
-	}
-
-	public void deletaClinica() {
-
-	}
-
-	public void alteraAvaliacao() {
-
-	}
-
-	public void alteraEndereco() {
-
-	}
-
-	public void adicionaMedico() {
-
-	}
-
-	public void removeMedico() {
-
-	}
-
-	public void alteraCrmv() {
-
-	}
-
-	public Medico getMedicoById() {
-		return null;
-	}
-
-	public List<Medico> getAllMedico() {
-		return null;
-	}
-
-}
