--- app/src/main/java/br/ufrpe/bsi/mpoo/petSpeed/negocio/EnderecoServices.java
@@ -1,34 +1,0 @@
-package br.ufrpe.bsi.mpoo.petSpeed.negocio;
-
-import br.ufrpe.bsi.mpoo.petSpeed.dominio.Endereco;
-import br.ufrpe.bsi.mpoo.petSpeed.infra.persistencia.EnderecoDAO;
-
-public class EnderecoServices {
-
-	private EnderecoDAO enderecoDAO;
-
-	public void cadastraEndereco() {
-
-	}
-
-	public void deletaEndereco() {
-
-	}
-
-	public void alteraCep() {
-
-	}
-
-	public void alteraNumero() {
-
-	}
-
-	public void alteraComplemento() {
-
-	}
-
-	public Endereco getEnderecoById() {
-		return null;
-	}
-
-}
