--- app/src/main/java/br/ufrpe/bsi/mpoo/petSpeed/negocio/MedicoServices.java
@@ -1,48 +1,0 @@
-package br.ufrpe.bsi.mpoo.petSpeed.negocio;
-
-import br.ufrpe.bsi.mpoo.petSpeed.dominio.Endereco;
-import br.ufrpe.bsi.mpoo.petSpeed.dominio.Medico;
-import br.ufrpe.bsi.mpoo.petSpeed.infra.persistencia.ClinicaDAO;
-import br.ufrpe.bsi.mpoo.petSpeed.infra.persistencia.EnderecoDAO;
-import br.ufrpe.bsi.mpoo.petSpeed.infra.persistencia.MedicoDAO;
-
-public class MedicoServices {
-
-	private MedicoDAO medicoDAO;
-
-	private EnderecoDAO enderecoDAO;
-
-	private ClinicaDAO clinicaDAO;
-
-
-	public void cadastraMedico(Medico medico) {
-
-		
-
-	}
-
-	public void deletaMedico() {
-
-	}
-
-	public void alteraAvaliacao() {
-
-	}
-
-	public void alteraClinica() {
-
-	}
-
-	public void alteraEndereco() {
-
-	}
-
-	public void alteraCrmv() {
-
-	}
-
-	public Endereco getEnderecoById() {
-		return null;
-	}
-
-}
