--- app/src/main/java/br/ufrpe/bsi/mpoo/petSpeed/negocio/PessoaServices.java
@@ -1,26 +1,0 @@
-package br.ufrpe.bsi.mpoo.petSpeed.negocio;
-
-
-import br.ufrpe.bsi.mpoo.petSpeed.infra.persistencia.PessoaDAO;
-
-public class PessoaServices {
-
-	private PessoaDAO pessoaDAO;
-
-	private boolean validaCpf() {
-		return false;
-	}
-
-	public void cadastraPessoa() {
-
-	}
-
-	public void deletaPessoa() {
-
-	}
-
-	public void alteraCpf() {
-
-	}
-
-}
