--- app/src/main/java/br/ufrpe/bsi/mpoo/petSpeed/negocio/TriagemServices.java
@@ -1,5 +1,0 @@
-package br.ufrpe.bsi.mpoo.petSpeed.negocio;
-
-public class TriagemServices {
-
-}
