--- app/src/main/java/br/ufrpe/bsi/mpoo/petSpeed/negocio/slopeOne.java
@@ -1,9 +1,0 @@
-package br.ufrpe.bsi.mpoo.petSpeed.negocio;
-
-public class slopeOne {
-
-	public void calculaRecomendaçao() {
-
-	}
-
-}
