+++ app/src/main/java/br/ufrpe/bsi/mpoo/petSpeed/os/dominio/Prioridade.java
@@ -1,0 +1,9 @@
+package br.ufrpe.bsi.mpoo.petSpeed.os.dominio;
+
+public enum Prioridade{
+
+	ALTA,
+
+	BAIXA
+
+}
