+++ app/src/main/java/br/ufrpe/bsi/mpoo/petSpeed/pessoa/dominio/Pessoa.java
@@ -1,0 +1,36 @@
+package br.ufrpe.bsi.mpoo.petSpeed.pessoa.dominio;
+
+public class Pessoa {
+
+	private long id;
+
+	private String nome;
+
+	private String cpf;
+
+
+
+	public long getId() {
+		return id;
+	}
+
+	public void setId(long id) {
+		this.id = id;
+	}
+
+	public String getNome() {
+		return nome;
+	}
+
+	public void setNome(String nome) {
+		this.nome = nome;
+	}
+
+	public String getCpf() {
+		return cpf;
+	}
+
+	public void setCpf(String cpf) {
+		this.cpf = cpf;
+	}
+}
