+++ app/src/main/java/br/ufrpe/bsi/mpoo/petSpeed/pessoa/negocio/EnderecoServices.java
@@ -1,0 +1,34 @@
+package br.ufrpe.bsi.mpoo.petSpeed.pessoa.negocio;
+
+import br.ufrpe.bsi.mpoo.petSpeed.pessoa.dominio.Endereco;
+import br.ufrpe.bsi.mpoo.petSpeed.pessoa.persistencia.EnderecoDAO;
+
+public class EnderecoServices {
+
+	private EnderecoDAO enderecoDAO;
+
+	public void cadastraEndereco() {
+
+	}
+
+	public void deletaEndereco() {
+
+	}
+
+	public void alteraCep() {
+
+	}
+
+	public void alteraNumero() {
+
+	}
+
+	public void alteraComplemento() {
+
+	}
+
+	public Endereco getEnderecoById() {
+		return null;
+	}
+
+}
