+++ app/src/main/java/br/ufrpe/bsi/mpoo/petSpeed/pessoa/negocio/PessoaServices.java
@@ -1,0 +1,26 @@
+package br.ufrpe.bsi.mpoo.petSpeed.pessoa.negocio;
+
+
+import br.ufrpe.bsi.mpoo.petSpeed.pessoa.persistencia.PessoaDAO;
+
+public class PessoaServices {
+
+	private PessoaDAO pessoaDAO;
+
+	private boolean validaCpf() {
+		return false;
+	}
+
+	public void cadastraPessoa() {
+
+	}
+
+	public void deletaPessoa() {
+
+	}
+
+	public void alteraCpf() {
+
+	}
+
+}
