+++ app/src/main/java/br/ufrpe/bsi/mpoo/petSpeed/pessoa/persistencia/EnderecoDAO.java
@@ -1,0 +1,35 @@
+package br.ufrpe.bsi.mpoo.petSpeed.pessoa.persistencia;
+
+
+import br.ufrpe.bsi.mpoo.petSpeed.pessoa.dominio.Endereco;
+import br.ufrpe.bsi.mpoo.petSpeed.infra.DBHelper;
+
+public class EnderecoDAO {
+
+	private DBHelper db;
+
+	public void cadastraEndereco() {
+
+	}
+
+	public void deletaEndereco() {
+
+	}
+
+	public void alteraCep() {
+
+	}
+
+	public void alteraNumero() {
+
+	}
+
+	public void alteraComplemento() {
+
+	}
+
+	public Endereco getEnderecoById() {
+		return null;
+	}
+
+}
