+++ app/src/main/java/br/ufrpe/bsi/mpoo/petSpeed/pessoa/persistencia/PessoaDAO.java
@@ -1,0 +1,25 @@
+package br.ufrpe.bsi.mpoo.petSpeed.pessoa.persistencia;
+
+import br.ufrpe.bsi.mpoo.petSpeed.infra.DBHelper;
+
+public class PessoaDAO {
+
+	private DBHelper db;
+
+	public void cadastraPessoa() {
+
+	}
+
+	public void deletaPessoa() {
+
+	}
+
+	public void alteraNome() {
+
+	}
+
+	public void alteraCPF() {
+
+	}
+
+}
